# docs/schema/swagger.yaml
openapi: 3.0.1
info:
  title: Building OS API
  description: An ASP.NET Core API for Building OS
  version: v1
paths:
...
  /telemetries/hot:
    get:
      tags:
        - Telemetry
      summary: "[非推奨] 最新のテレメトリデータを取得（Hot 層 直接）。正本は `GET /telemetries/query?latest=true`。\r\nこの per-tier エンドポイントは後方互換のため残置。"
      parameters:
        - name: pointId
          in: query
          description: 必須. ポイントID
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ValidTelemetryData'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ValidTelemetryData'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ValidTelemetryData'
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      deprecated: true
  /telemetries/warm:
    get:
      tags:
        - Telemetry
      summary: '[非推奨] 指定期間の Warm テレメトリデータを取得（Warm 層 直接）。正本は `GET /telemetries/query`（tier 自動選択）。'
      parameters:
        - name: pointId
          in: query
          schema:
            type: string
        - name: startTime
          in: query
          schema:
            type: string
            format: date-time
        - name: endTime
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ValidTelemetryData'
            application/json:
              schema:
                $ref: '#/components/schemas/ValidTelemetryData'
            text/json:
              schema:
                $ref: '#/components/schemas/ValidTelemetryData'
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      deprecated: true
  /telemetries/cold:
    get:
      tags:
        - Telemetry
      summary: '[非推奨] 指定期間の Cold テレメトリデータを取得（Cold 層 直接）。正本は `GET /telemetries/query`（tier 自動選択）。'
      parameters:
        - name: pointId
          in: query
          schema:
            type: string
        - name: startTime
          in: query
          schema:
            type: string
            format: date-time
        - name: endTime
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ValidTelemetryData'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ValidTelemetryData'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ValidTelemetryData'
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      deprecated: true
  /telemetries/cold-multi-point:
    get:
      tags:
        - Telemetry
      summary: '[非推奨] 指定期間の Cold テレメトリデータを取得（複数ポイント, Cold 層 直接）。正本は `GET /telemetries/query`。'
      parameters:
        - name: pointIds
          in: query
          schema:
            type: array
            items:
              type: string
        - name: startTime
          in: query
          schema:
            type: string
            format: date-time
        - name: endTime
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: object
                additionalProperties:
                  type: array
                  items:
                    $ref: '#/components/schemas/ValidTelemetryData'
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: array
                  items:
                    $ref: '#/components/schemas/ValidTelemetryData'
            text/json:
              schema:
                type: object
                additionalProperties:
                  type: array
                  items:
                    $ref: '#/components/schemas/ValidTelemetryData'
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      deprecated: true
  /telemetries/query:
    get:
      tags:
        - Telemetry
      summary: "テレメトリ取得の正本エンドポイント。期間・粒度・latest を指定し、tier（hot/warm/cold/集計）を\r\n自動選択する。per-tier の `/hot`・`/warm`・`/cold`・`/cold-multi-point` は非推奨。"
      parameters:
        - name: pointId
          in: query
          description: 必須. ポイントID
          schema:
            type: string
        - name: start
          in: query
          description: 開始時刻（latest=true の場合は不要）
          schema:
            type: string
            format: date-time
        - name: end
          in: query
          description: 終了時刻（latest=true の場合は不要）
          schema:
            type: string
            format: date-time
        - name: granularity
          in: query
          description: '集計粒度: raw / hour / day（省略時: raw）'
          schema:
            $ref: '#/components/schemas/TelemetryGranularity'
        - name: latest
          in: query
          description: true の場合は最新値のみ返す
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ValidTelemetryData'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ValidTelemetryData'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ValidTelemetryData'
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
components:
  schemas:
...
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
      additionalProperties: { }
...
    TelemetryGranularity:
      enum:
        - 0
        - 1
        - 2
      type: integer
      format: int32
...
    ValidTelemetryData:
      type: object
      properties:
        building:
          type: string
          nullable: true
        data:
          type: string
          nullable: true
        datetime:
          type: string
          nullable: true
        deviceId:
          type: string
          nullable: true
        id:
          type: string
          nullable: true
        name:
          type: string
          nullable: true
        pointId:
          type: string
          nullable: true
        value:
          type: number
          format: double
          nullable: true
      additionalProperties: false